+++ docs/source/auto_examples/example_nn_Si.py
@@ -102,7 +102,7 @@ model.set_save_metadata(prefix="./kliff_saved_model", start=5, frequency=2)
 # :mod:`~kliff.calculators.CalculatorTorch()`, which is targeted for the NN model.
 # Also, its ``create()`` method takes an argument ``reuse`` to inform whether to reuse the
 # fingerprints generated from the descriptor if it is present.
-# To train on gpu, set ``gpu=True`` in ``Calculator``.
+# To train on gpu, set ``gpu=True`` in ``CalculatorNew``.
 #
 
 # training set