+++ docs/source/changelog.rst
@@ -91,7 +91,7 @@ v0.1.2 (2019/6/27)
 - various bug fixes
 - API changes
   * class ``DataSet`` renamed to ``Dataset``
-  * class ``Calculator`` moved to module ``calculators`` from module ``calculator``
+  * class ``CalculatorNew`` moved to module ``calculators`` from module ``calculator``
 
 
 v0.1.1 (2019/5/13)